moved event.ts into state/
