+++ dashi/src/lib/actions/handleComponentChange.ts
@@ -1,7 +1,7 @@
 import { store } from "@/lib/store";
 import { type ContribPoint } from "@/lib/types/model/extension";
 import { type CallbackRequest } from "@/lib/types/model/callback";
-import { type ComponentChangeEvent } from "@/lib/types/model/event";
+import { type ComponentChangeEvent } from "@/lib/types/state/event";
 import { getInputValues } from "@/lib/actions/helpers/getInputValues";
 import { applyStateChangeRequests } from "@/lib/actions/helpers/applyStateChangeRequests";
 import { invokeCallbacks } from "@/lib/actions/helpers/invokeCallbacks";
@@ -64,6 +64,7 @@ function getCallbackRequests(
           equalObjPaths(input.property, changeEvent.property),
       );
       if (inputIndex >= 0) {
+        // Collect triggered callback
         callbackRequests.push({
           contribPoint,
           contribIndex,
