+++ dashi/src/lib/components/Button.tsx
@@ -2,7 +2,7 @@ import { type MouseEvent } from "react";
 import MuiButton from "@mui/material/Button";
 
 import { type ButtonState } from "@/lib/types/state/component";
-import { type ComponentChangeHandler } from "@/lib/types/model/event";
+import { type ComponentChangeHandler } from "@/lib/types/state/event";
 
 export interface ButtonProps extends Omit<ButtonState, "type"> {
   onChange: ComponentChangeHandler;
