+++ dashi/src/lib/types/state/event.ts
